api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
notebooklm-py==0.1.1
python-dotenv==1.0.1
pydantic==2.10.4
